delete big file(logs ..)
Former-commit-id: 805cdea58dc08c7dc6baac38bb3151933dfcf78f
Former-commit-id: 699947a4813dae575925660855b49f6a43085705

+++ deploy_test_db/scp_iotdb.sh
@@ -7,15 +7,23 @@ dest_db_dir=`cat ${conf_file}|grep ^db_dir|awk -F '=' '{print $2}'`
 source_db_dir=`cat ${conf_file}|grep ^source_db_dir|awk -F '=' '{print $2}'`
 
 # unzip local 
-unzip ${source_db_dir}/*.zip -d ${source_db_dir}/
-mv ${source_db_dir}/*-bin/* ${source_db_dir}/
-cp -rp ${source_db_dir}/conf ${source_db_dir}/conf_orig
 
 if [ -f "${source_db_dir}/sbin/start-cli.sh" ]; then
     echo "success"
 else
+    unzip ${source_db_dir}/*.zip -d ${source_db_dir}/
+    mv ${source_db_dir}/*-bin/* ${source_db_dir}/
+    cp -rp ${source_db_dir}/conf ${source_db_dir}/conf_orig
+
     return 
 fi
+
+if [ -f "${source_db_dir}/conf_orig" ]; then
+    echo "success"
+else
+    cp -rp ${source_db_dir}/conf ${source_db_dir}/conf_orig 
+fi
+
 
 exec 3<${nodeinfo_dir}/datanode.txt
 while read line<&3
